+++ packages/react/src/index.ts
@@ -18,3 +18,4 @@ export * from "@nimbus-ds/sidemodal";
 export * from "@nimbus-ds/thumbnail-with-action";
 export * from "@nimbus-ds/calendar";
 export * from "@nimbus-ds/empty-app";
+export * from "@nimbus-ds/initial-screen";